FILE: moves getAdversariesById
