+++ libs/spirit-islander/util/src/lib/game/get-adversary-by-id.spec.ts
